Split context module into submodules
